% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arr_vars.R
\name{arr_vars}
\alias{arr_vars}
\title{arr_vars: Arrange columns}
\usage{
arr_vars(data, vars)
}
\arguments{
\item{data}{Dataframe class object}

\item{vars}{character vector that identifies position of each variable}
}
\value{
Dataframe class object
}
\description{
Rearranges columns in a dataframe
}
\details{
Useful for reading / displaying datasets
}
\examples{
data(iris)
head(arr_vars(iris, c('Species'=1)))

}
\author{
Kevin W. McConeghy
}
