% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange.vars.R
\name{arrange.vars}
\alias{arrange.vars}
\title{arrange.vars: Arrange columns}
\usage{
\method{arrange}{vars}(data, vars)
}
\arguments{
\item{data}{Dataframe class object}

\item{vars}{character vector that identifies position of each variable}
}
\value{
Dataframe class object
}
\description{
Rearranges columns in a dataframe
}
\details{
Useful for reading / displaying datasets
}
\examples{
data(iris)
Put Species as first column
head(arrange.vars(iris, c('Species'=1)))

}
\author{
Kevin W. McConeghy
}
