% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Boot.ATE.R
\name{Boot.ATE}
\alias{Boot.ATE}
\title{Boot.ATE: Bootstrapped marginal effects}
\usage{
Boot.ATE(model, treat, R = 250, block = "", df)
}
\arguments{
\item{model}{GLM regression object}

\item{treat}{Character string for variable to calculate ATE with}

\item{R}{Integer for number of bootstrap replications}

\item{block}{Character string for "block" to sample by (cluster robust SE's).}

\item{df}{Dataframe object of original sample}
}
\value{
coefficient from glm object

coefficient from bootstraps

ATE (p1- p0)

RR (p1/p0)

matrix of bootstrap replications
}
\description{
Estimate Average Treatment Effect with Bootstrapped CI's
}
\details{
Boot.ATE will take a glm regression object and compute predicted outcomes with
treatment = 1 and treatment 0; It will then bootstrap the model and compute effects.
There is an option for block bootstrapping to compute cluster robust intervals.
Bootstrapped confidence intervals are percentile.
}
\examples{
require(geepack)
require(boot)
data(ohio)
model <- glm(resp ~ age + smoke, data=ohio, family=binomial())
b1 <- Boot.ATE(model, treat="smoke", R=10000, df=ohio, block="id", type="perc")
b1$ate

b1$coeff #Bootstrapped
confint.geeglm <- function(object, parm, level = 0.95, ...) {
 cc <- coef(summary(object))
 mult <- qnorm((1+level)/2)
 citab <- with(as.data.frame(cc),
               cbind(lwr=Estimate-mult*Std.err,
                     upr=Estimate+mult*Std.err))
 rownames(citab) <- rownames(cc)
 citab[parm,]
 }
confint.geeglm(g1, "smoke") #Comparison to GEE
}
\author{
Kevin W. McConeghy
}
