% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_ATE.R
\name{boot_ATE}
\alias{boot_ATE}
\title{boot_ATE: Bootstrapped marginal effects}
\usage{
boot_ATE(model, treat, R = 250, block = "", df)
}
\arguments{
\item{model}{GLM regression object}

\item{treat}{Character string for variable to calculate ATE with}

\item{R}{Integer for number of bootstrap replications}

\item{block}{Character string for "block" to sample by (cluster robust SE's).}

\item{df}{Dataframe object of original sample}
}
\value{
coefficient from glm object

coefficient from bootstraps

ATE (p1- p0)

RR (p1/p0)

matrix of bootstrap replications
}
\description{
Estimate Average Treatment Effect with Bootstrapped CI's
}
\details{
boot_ATE will take a glm regression object and compute predicted outcomes with
treatment = 1 and treatment 0; It will then bootstrap the model and compute effects.
There is an option for block bootstrapping to compute cluster robust intervals.
Bootstrapped confidence intervals are percentile.
}
\author{
Kevin W. McConeghy
}
