% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/des_char.R
\name{des_char}
\alias{des_char}
\title{des_char}
\usage{
des_char(fvar)
}
\arguments{
\item{fvar}{string variable}
}
\value{
Object with generic information, statistics
}
\description{
Exploratory data analysis of a string character var
}
\details{
This function will take a dataframe variable, provide summary and graphs
}
\examples{
data("acq")

}
\author{
Kevin W. McConeghy
}
