% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/des_num.R
\name{des_num}
\alias{des_num}
\title{des_num}
\usage{
des_num(numvar)
}
\arguments{
\item{numvar}{dataframe object}
}
\value{
Object with generical information, statistics, ggplot2 plots
}
\description{
Exploratory data analysis of a numeric variable
}
\details{
This function will take a dataframe, provide summary statistics, distributions checks, graphs
}
\note{
if no filepath specified, will default to working directory with timestamp.
}
\examples{

}
\author{
Kevin W. McConeghy
}
