% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/des_var.R
\name{des_var}
\alias{des_var}
\title{des_var}
\usage{
des_var(x)
}
\arguments{
\item{x}{object}
}
\value{
List with HtmlTable, var statistics
}
\description{
Generate an object with html table of var statistics
}
\details{
This function will take a variable, provide summary statistics, distributions checks, graphs
}
\note{
Excepts var class integer, numeric
}
\examples{

}
\author{
Kevin W. McConeghy
}
