% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DfDes.R
\name{DfDes}
\alias{DfDes}
\title{DfDes: Simple description of dataframe}
\usage{
DfDes(df, dfName = "", VarList = NA, CheckMiss = T)
}
\arguments{
\item{df}{dataframe class object}

\item{dfName}{character vector of length 1, title of dataset}

\item{VarList}{character vector of variable names which will be checked for unique values}

\item{CheckMiss}{default=True, will report number of NA values}
}
\value{
Prints report
}
\description{
Reports on dataset size, row counts, unique variables. Requires dplyr.
}
\examples{
data(iris)
DfDes(iris, 'Flower dataset', c('Petal.Length'))

}
\author{
Kevin W. McConeghy
}
