% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FactorEDA.R
\name{FactorEDA}
\alias{FactorEDA}
\title{FactorEDA}
\usage{
FactorEDA(fvar)
}
\arguments{
\item{fvar}{dataframe object}
}
\value{
Object with generical information, statistics, ggplot2 plots
}
\description{
Exploratory data analysis of a Factor variable
}
\details{
This function will take a dataframe, provide summary statistics graphs
}
\note{
if no filepath specified, will default to working directory with timestamp.
}
\examples{

}
\author{
Kevin W. McConeghy
}
