% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_loess.R
\name{gg_loess}
\alias{gg_loess}
\title{gg_loess: Loess plot for propensity scores}
\usage{
gg_loess(x, response, treatment, responseTitle = "",
  treatmentTitle = "Treatment", percentPoints.treat = 0.2,
  percentPoints.control = 0.1, points.treat.alpha = 0.2,
  points.control.alpha = 0.1, ...)
}
\arguments{
\item{x}{vector of propensity scores.}

\item{response}{the response variable.}

\item{treatment}{the treatment varaible as a logical type.}

\item{responseTitle}{the label to use for the y-axis (i.e. the name of the response variable)}

\item{treatmentTitle}{the label to use for the treatment legend.}

\item{percentPoints.treat}{the percentage of treatment points to randomly plot.}

\item{percentPoints.control}{the percentage of control points to randomly plot.}

\item{points.treat.alpha}{the transparency level for treatment points.}

\item{points.control.alpha}{the transparency level for control points.}

\item{...}{other parameters passed to \code{\link{geom_smooth}} and
\code{\link{stat_smooth}}.}
}
\value{
a ggplot2 figure
}
\description{
Loess plot with density distributions for propensity scores and outcomes on
top and right, respectively.
}
\examples{
require(MatchIt)
lalonde <- lalonde
m.out <- matchit(treat ~ age + educ + black + hispan, data=lalonde, method="nearest")
df_m <- match.data(m.out)
gg_loess(df_m$distance, df_m$age, df_m$treat)


}
