% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.packages.R
\name{load.packages}
\alias{load.packages}
\title{Check for package and load if found}
\usage{
load.packages(packages, quietly = F)
}
\arguments{
\item{packages}{vector list of packages to install}
}
\description{
Check for package and load if found
}
\examples{
load.packages("plyr")

}
