% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.project.R
\name{load.project}
\alias{load.project}
\title{load.project}
\usage{
load.project(FilePath, INSHEET = FALSE, MUNGE = FALSE, ANALYZE = FALSE,
  REPORT = FALSE)
}
\arguments{
\item{FilePath}{Root folder where insheet etc. located}

\item{INSHEET}{logic indicator for which files to run}

\item{MUNGE}{logic indicator for which files to run}

\item{ANALYZE}{logic indicator for which files to run}

\item{REPORT}{logic indicator for which files to run}
}
\description{
Initiate Project Commands
}
\examples{

FilePath <- 'mydir'
load.project(FilePath,INSHEET=TRUE) #Run Insheet only


}
