% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric_funcs.R
\name{num_cont}
\alias{num_cont}
\title{num_cont: Compute mean (standard deviation)}
\usage{
numeric_cont(x, .digits=2)
}
\arguments{
\item{x}{A numeric vector}

\item{.digits}{Significant figures to round to}
}
\value{
A character string of length 1, 'mean (SD)'
}
\description{
Given a single numeric vector, will compute mean and
standard deviation, returning results in string. Mostly called by
other functions.
}
\examples{
require(Scotty)
num_cont(mtcars$mpg)

}
