% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric_funcs.R
\name{num_dt}
\alias{num_dt}
\title{num_dt: Compute median date, range.}
\usage{
num_dt(x, median = TRUE, range='min/max', origin=NULL)
}
\arguments{
\item{range}{A character string to indicate range values. options:
'min/max'.}

\item{origin}{Default NULL. If specified,
must be a date class object of length 1. For use if date variable is in
raw integer format.}

\item{x}{A integer or date class vector}

\item{nomedian}{A logical vector, indicates whether to report median or not.
Default = TRUE.}
}
\value{
A character string of length 1, 'Median date (date1, date2)'
}
\description{
Given a single integer vector (with origin specified)
or data class vector, will compute median day and a range
(default min/max) and percent of total, returning results in string.
Mostly called by other functions.
}
\examples{
require(mdsR)
mds_dta <- mdsR::mds_cohort
num_dt(mds_dta$dmdate)

}
