% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric_funcs.R
\name{num_fctr}
\alias{num_fctr}
\title{num_fctr: Compute number (proportion as percent)}
\usage{
num_fctr(x, .digits=2, .referent = TRUE)
}
\arguments{
\item{x}{A numeric or factor class vector}

\item{.digits}{Significant figures to round to}

\item{.referent}{Value to compute no. (percent) for}
}
\value{
A character string of length 1, 'No. (percent)'
}
\description{
Given a single numeric vector or factor,
will compute no. equal to a .referent value and percent of total,
returning results in string. Mostly called by other functions.
}
\examples{
require(Scotty)
num_fctr(mtcars$cyl, .referent=4)

}
