% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_one.R
\name{render_one}
\alias{render_one}
\title{render_one: Render a single .Rmd file, and output to location}
\usage{
render_one(CodeFileId, SrcPath, HtmlPath, stamp = T)
}
\arguments{
\item{CodeFileId}{file prefix with Scotty canon name, e.g. 'A01'}

\item{SrcPath}{filepath where '.Rmd' files exist}

\item{stamp}{default=True, will fix date-time to file name.}

\item{HtmPath}{filepath to render to}
}
\description{
render_one: Render a single .Rmd file, and output to location
}
\examples{
Scotty::render_one('A01', 'CodeFilesPath', paste0('ReportFilesPath'))

}
\author{
Kevin W. McConeghy
}
