% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t1_binvar.R
\name{t1_binvar}
\alias{t1_binvar}
\title{t1_binvar: "Table One", Binary Variable Summary}
\usage{
t1_binvar(data, .var, .group, .referent = "1", quietly = T)
}
\arguments{
\item{.var}{Variable to be summarized}

\item{.group}{Group to summarize over}

\item{.referent}{If variable logical or numeric, default reported is 1/"T",
otherwise specify here.}

\item{quietly}{If FALSE, function will report some internal steps}

\item{df}{Dataframe}
}
\value{
List of length = groups
}
\description{
Function which summarizes dichotomous or binary
variable by a named group
}
\details{
Supply a dataframe, a variable to summarize, a group variable
}
\examples{

}
\author{
Kevin W. McConeghy
}
