% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t1_normvar.R
\name{t1_normvar}
\alias{t1_normvar}
\title{t1_binvar: "Table One", Binary Variable Summary}
\usage{
t1_normvar(data, .var, .group, quietly = T)
}
\arguments{
\item{data}{Dataframe}

\item{.var}{Variable to be summarized}

\item{.group}{Group to summarize over}

\item{quietly}{If FALSE, function will report some internal steps}
}
\value{
List of length = groups
}
\description{
Function which summarizes dichotomous or binary
variable by a named group
}
\details{
Supply a dataframe, a variable to summarize, a group variable
}
\examples{

}
\author{
Kevin W. McConeghy
}
