% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_files.R
\name{tab_files}
\alias{tab_files}
\title{tab_files}
\usage{
tab_files(files)
}
\description{
Helper function, takes Scotty canon list of files and
outputs an HTML DT table, for use with rmarkdown.
}
\examples{
tab_files()

}
