% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarByGr.R
\name{VarByGr}
\alias{VarByGr}
\title{VarByGr}
\usage{
VarByGr(df, var, group, VarTypes)
}
\arguments{
\item{df}{Dataframe}

\item{var}{Variable to be summarized}

\item{group}{Group to summarize over}

\item{VarTypes}{list with three elements, normvars nonnormvars binvars}
}
\value{
List of length = groups
}
\description{
Function which summarizes numeric/int by group
}
\details{
First object column 1 (group\link{1}), second obj column 2 (group\link{2})
to be used in generating a "table one"
}
\examples{

}
\author{
Kevin W. McConeghy
}
