% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsx_Break.R
\name{xlsx_break}
\alias{xlsx_break}
\title{xlsx_break}
\usage{
xlsx_break(sheet, numberOfLine = 1)
}
\arguments{
\item{sheet}{A sheet class object created by xlsx package (see example)}

\item{numberOfLine}{Number of linebreaks}
}
\description{
Several commands to be used with xlsx package
}
\examples{

}
\author{
Kevin W. McConeghy
}
