% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsx_header.R
\name{xlsx_header}
\alias{xlsx_header}
\title{xlsx_header}
\usage{
xlsx_header(wb, sheet, value = "Header", StartRow = NULL, StartCol = 1)
}
\arguments{
\item{wb}{Workbook object}

\item{sheet}{A sheet class object created by xlsx package (see example)}

\item{value}{Input to be placed into cell}

\item{StartRow}{excel file starting row number}

\item{StartCol}{excel file starting col number}
}
\description{
Several commands to be used with xlsx package
}
\details{
Package contains xlsx.AddHeader,
}
\examples{

}
\author{
Kevin W. McConeghy
}
