% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsx_paragph.R
\name{xlsx_paragph}
\alias{xlsx_paragph}
\title{xlsx_paragph}
\usage{
xlsx_paragph(wb, sheet, value, fontColor = "#FFFFFF", fontSize = 12,
  backGroundColor = "#FFFFFF", isBold = FALSE, isItalic = FALSE,
  startRow = NULL, startCol = 1, colSpan = 10, rowSpan = 6)
}
\arguments{
\item{wb}{Workbook object}

\item{sheet}{A sheet class object created by xlsx package (see example)}

\item{value}{Input to be placed into cell}

\item{fontColor}{Color font}

\item{fontSize}{Excel font size}

\item{backGroundColor}{Excel background color}

\item{isBold}{logical indicating if font is bold or not}

\item{isItalic}{logical indicating italic font}

\item{startRow}{integer indicating excel row to insert}

\item{startCol}{integer indicating excel column to insert}

\item{colSpan}{Number of columns to paste across}

\item{rowSpan}{Number of rows to paste down}
}
\description{
Several commands to be used with xlsx package
}
\details{
Package contains xlsx_paragph,
}
\examples{

}
\author{
Kevin W. McConeghy
}
