% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsx_plot.R
\name{xlsx_plot}
\alias{xlsx_plot}
\title{xlsx_plot}
\usage{
xlsx_plot(wb, sheet, plotFunction, startRow = NULL, startCol = 1,
  width = 480, height = 480, ...)
}
\arguments{
\item{wb}{Workbook class object}

\item{sheet}{Sheet class object}

\item{plotFunction}{Plot function to execute with return value plot}

\item{startRow}{integer for row to insert at}

\item{startCol}{integer for column to insert at}

\item{width}{Figure width}

\item{height}{Figure height}

\item{...}{other arguments to pass on to png()}
}
\description{
Several commands to be used with xlsx package
}
\examples{

}
\author{
Kevin W. McConeghy
}
