% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsx_table.R
\name{xlsx_table}
\alias{xlsx_table}
\title{xlsx_table}
\usage{
xlsx_table(wb, sheet, data, startRow = NULL, startCol = 1,
  col.names = TRUE, row.names = TRUE, columnWidth = 14,
  fontColor = "#FFFFFF", fontSize = 12, rownamesFill = "white",
  colnamesFill = "white", rowFill = c("white", "white"))
}
\arguments{
\item{wb}{Workbook object}

\item{sheet}{A sheet class object created by xlsx package (see example)}

\item{data}{Dataframe class object}

\item{startRow}{excel file starting row number (top left of table)}

\item{startCol}{excel file starting col number (top left of table)}

\item{col.names}{logical for using column names}

\item{row.names}{logical for using row names}

\item{columnWidth}{Width of column cells}

\item{fontColor}{Excel font color}

\item{fontSize}{Excel font size}

\item{rownamesFill}{color for font background}

\item{colnamesFill}{color for font background}

\item{rowFill}{color for cell background in rows}
}
\description{
Several commands to be used with xlsx package
}
\details{
Package contains xlsx_table,
}
\examples{

}
\author{
Kevin W. McConeghy
}
