% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_pkgs.R
\name{load_pkgs}
\alias{load_pkgs}
\title{load_pkgs}
\usage{
load_pkgs()
}
\arguments{
\item{packages}{vector list of packages to install}
}
\description{
Check for package and load if found
}
\examples{
load_pkgs(c("dplyr", "ggplot2", "tm")

}
