% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_prj.R
\name{load_prj}
\alias{load_prj}
\title{load_prj}
\usage{
load_prj(INSHEET = F, MUNGE = F, REPORT = F, SrcPath, ReportPath)
}
\arguments{
\item{INSHEET}{logical for which files to run}

\item{MUNGE}{logical for which files to run}

\item{REPORT}{logical for which files to run}

\item{SrcPath}{Path whether markdown files exist}

\item{OutPath}{Path to place rendered files}
}
\description{
Initiate project commands in Scotty canon project, this program will search the SrcPath
directory for files which conform to the nomenclature, 'A\link{0-9}', 'B\link{0-9}',
'C\link{0-9}', 'D\link{0-9}'
}
\examples{

FilePath <- 'mydir'
CodePath <- 'CodeDir'
load_prj(INSHEET=TRUE, SrcPath=CodePath, ReportPath = FilePath) #Run Insheet only


}
